Terminal has essentials not nice to haves tba.
As of now the bare bones of the system are in place. I'd say I would be
happy leaving it as is if with the caveat being it is only useful to me
in the context of two stops with only the T being including in it. What
I need now is for the program to generate a route from adresses and show
alerts in the system. Delays don't show up and can be confusing for
users. Additionally, we should add the MBTA recommended smoothing for
time. Finally I want bus route options available.